solutions/flask-products/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-CORS>=4.0.0
Flask-RESTful>=0.3.10
Marshmallow>=3.19.0
flask-marshmallow>=0.15.0
marshmallow-sqlalchemy>=0.29.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-flask>=1.2.0
mcp>=1.0.0